change mentsu.js to mentsu.ts
