+++ src/test/js/jan/hand-util-test.js
@@ -21,7 +21,7 @@ import Hand from '../../../main/js/jan/hand';
 import HandUtil from '../../../main/js/jan/hand-util';
 import {Janpai} from '../../../main/js/jan/janpai';
 import {JANPAI_ID} from '../../../main/js/jan/janpai-id';
-import Mentsu from '../../../main/js/jan/mentsu';
+import {Mentsu} from '../../../main/js/jan/mentsu';
 
 describe('HandUtilTest', () => {
