+++ src/test/js/jan/jan-util-test.js
@@ -19,7 +19,7 @@ import Hand from '../../../main/js/jan/hand';
 import {Janpai} from '../../../main/js/jan/janpai';
 import {JANPAI_ID} from '../../../main/js/jan/janpai-id';
 import JanUtil from '../../../main/js/jan/jan-util';
-import Mentsu from '../../../main/js/jan/mentsu';
+import {Mentsu} from '../../../main/js/jan/mentsu';
 
 describe('JanUtilTest', () => {
